+++ ui/src/audit_registry.rs
@@ -158,6 +158,7 @@ pub enum AuditId {
     CenterWt2d,
     CenterWt2dToolbar,
     CenterWt2dPlot,
+    CenterWt2dResult,
     CenterWt2dCurveEditor,
     CenterWt2dShapeEditor,
     CenterWt2dAnalyzeDialog,
@@ -266,7 +267,7 @@ impl AuditId {
             | Self::CenterScopeCellOsc | Self::CenterScopeCellFilter
             | Self::CenterScopeCellFx | Self::CenterScopeCellOut | Self::RailFilterTabs
             | Self::RailFiltEnvGraph | Self::RailAmpEnvGraph | Self::CenterWt2d
-            | Self::CenterWt2dPlot | Self::CenterWt2dCurveEditor | Self::CenterWt2dShapeEditor
+            | Self::CenterWt2dPlot | Self::CenterWt2dResult | Self::CenterWt2dCurveEditor | Self::CenterWt2dShapeEditor
             | Self::CenterWt3dStack | Self::CenterWt3dMorph | Self::FooterPianoCompact
             | Self::RailLevelMeter | Self::WidgetKnobSm | Self::WidgetKnobMd | Self::WidgetKnobLg
             | Self::WidgetPanel => AuditChecks::B,
@@ -340,6 +341,7 @@ fn static_label(id: &AuditId) -> &'static str {
         AuditId::CenterWt2d => "center.wt_2d",
         AuditId::CenterWt2dToolbar => "center.wt_2d.toolbar",
         AuditId::CenterWt2dPlot => "center.wt_2d.plot",
+        AuditId::CenterWt2dResult => "center.wt_2d.result",
         AuditId::CenterWt2dCurveEditor => "center.wt_2d.curve_editor",
         AuditId::CenterWt2dShapeEditor => "center.wt_2d.shape_editor",
         AuditId::CenterWt2dAnalyzeDialog => "center.wt_2d.analyze_dialog",
@@ -622,6 +624,7 @@ fn parent_bounds(
         AuditId::CenterWtStripCell(_) | AuditId::CenterWtStripLayerChip(_) => Some(AuditId::CenterWtStrip),
         AuditId::CenterWt2dToolbar
         | AuditId::CenterWt2dPlot
+        | AuditId::CenterWt2dResult
         | AuditId::CenterWt2dCurveEditor
         | AuditId::CenterWt2dShapeEditor => Some(AuditId::CenterWt2d),
         AuditId::RailFilterTabs | AuditId::RailFilterKnobs => Some(AuditId::RailPanelFilter),